리뷰 코멘트에 맞춰 local instance 경계를 정리한다
