+++ apps/api/src/graphql/resolvers/profile/mutation/create.ts
@@ -1,13 +1,7 @@
-import {
-  AccountProfiles,
-  db,
-  firstOrThrow,
-  isUniqueViolation,
-  Profiles,
-  resolveConfiguredLocalInstance,
-} from '@kosmo/core/db';
+import { AccountProfiles, db, firstOrThrow, isUniqueViolation, Profiles } from '@kosmo/core/db';
 import { AccountProfileRole, ProfileFollowPolicy } from '@kosmo/core/enums';
 import { ConflictError } from '@kosmo/core/error';
+import { resolveConfiguredLocalInstance } from '@kosmo/core/local-instance';
 import { normalizeHandle } from '@kosmo/core/utils';
 import { profileHandleSchema } from '@kosmo/core/validation';
 import { builder } from '@/graphql/builder';
