+++ apps/api/src/index.ts
@@ -1,9 +1,13 @@
 import { serve } from '@hono/node-server';
+import { db } from '@kosmo/core/db';
+import { resolveConfiguredLocalInstance } from '@kosmo/core/local-instance';
 import { Hono } from 'hono';
 import { deriveContext } from './context';
 import { yoga } from './graphql';
 import { rest } from './rest';
 import type { Env } from './context';
+
+await resolveConfiguredLocalInstance(db);
 
 const app = new Hono<Env>();
