+++ apps/web/e2e/global-setup.ts
@@ -2,7 +2,8 @@ export default async function globalSetup() {
   process.env.DATABASE_URL ??= 'postgres://kosmo:kosmo@localhost:54329/kosmo_test';
   process.env.PUBLIC_ORIGIN ??= 'http://127.0.0.1:4173';
 
-  const { bootstrapConfiguredLocalInstance, db, pg } = await import('@kosmo/core/db');
+  const { db, pg } = await import('@kosmo/core/db');
+  const { bootstrapConfiguredLocalInstance } = await import('@kosmo/core/db/seed');
 
   await pg.unsafe(`
     DO $$
