+++ packages/core/db/index.ts
@@ -5,7 +5,6 @@ import { relations } from './relations';
 import * as tables from './tables';
 
 export * from './id';
-export * from './local-instance';
 export * from './relations';
 export * from './tables';
 export * from './utils';
