+++ packages/core/db/seed/index.ts
@@ -1,0 +1,1 @@
+export * from './local-instance';
