+++ packages/core/package.json
@@ -5,9 +5,11 @@
     ".": "./index.ts",
     "./datetime": "./datetime.ts",
     "./db": "./db/index.ts",
+    "./db/seed": "./db/seed/index.ts",
     "./env": "./env.ts",
     "./error": "./error/index.ts",
     "./enums": "./enums.ts",
+    "./local-instance": "./local-instance.ts",
     "./polyfill": "./polyfill.ts",
     "./search": "./search.ts",
     "./tiptap": "./tiptap/index.ts",
